refactor: use BigDecimal for monetary values and update currency conversion entities and database schema

+++ src/main/java/com/karthick/expenz/currency/dto/CurrencyConversionRateDTO.java
@@ -1,7 +1,9 @@
 package com.karthick.expenz.currency.dto;
 
+import java.math.BigDecimal;
+
 public record CurrencyConversionRateDTO(
   String fromCurrency,
   String toCurrency,
-  Double rate
+  BigDecimal rate
 ) {}
