+++ src/main/java/com/karthick/expenz/expenses/dto/ExpenseGroupDTO.java
@@ -1,16 +1,17 @@
 package com.karthick.expenz.expenses.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ExpenseGroupDTO(
-  long id,
+  Long id,
   String title,
   String description,
-  long totalExpensesCount,
-  long totalIncomesCount,
-  double totalExpensesAmount,
-  double totalIncomesAmount,
-  double balanceAmount,
+  Long totalExpensesCount,
+  Long totalIncomesCount,
+  BigDecimal totalExpensesAmount,
+  BigDecimal totalIncomesAmount,
+  BigDecimal balanceAmount,
   List<ExpenseDTO> expenses,
   List<PieDataItem> expensePieDataItems,
   List<PieDataItem> incomePieDataItems
